# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "义乌商贸库存管家",
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/v1/transfers": {
      "get": {
        "summary": "Transfers",
        "operationId": "transfers_api_v1_transfers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Transfer",
        "operationId": "create_transfer_api_v1_transfers_post",
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 120,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Transfer"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transfers/{tid}/receive": {
      "post": {
        "summary": "Receive",
        "operationId": "receive_api_v1_transfers__tid__receive_post",
        "parameters": [
          {
            "name": "tid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tid"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 120,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Receive"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transfers/{tid}/discrepancy": {
      "post": {
        "summary": "Discrepancy",
        "operationId": "discrepancy_api_v1_transfers__tid__discrepancy_post",
        "parameters": [
          {
            "name": "tid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tid"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 120,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Close"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transfer-claims/{cid}/review": {
      "post": {
        "summary": "Claim Review",
        "operationId": "claim_review_api_v1_transfer_claims__cid__review_post",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cid"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 120,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Review"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transfers/{tid}/{action}": {
      "post": {
        "summary": "Transfer Action",
        "operationId": "transfer_action_api_v1_transfers__tid___action__post",
        "parameters": [
          {
            "name": "tid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tid"
            }
          },
          {
            "name": "action",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "reserve",
                "ship",
                "cancel"
              ],
              "type": "string",
              "title": "Action"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 120,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferAction"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/returns": {
      "get": {
        "summary": "Returns",
        "operationId": "returns_api_v1_returns_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Return Create",
        "operationId": "return_create_api_v1_returns_post",
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 120,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/CustomerReturn"
                  },
                  {
                    "$ref": "#/components/schemas/SupplierReturn"
                  }
                ],
                "discriminator": {
                  "propertyName": "kind",
                  "mapping": {
                    "customer": "#/components/schemas/CustomerReturn",
                    "supplier": "#/components/schemas/SupplierReturn"
                  }
                },
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quality": {
      "get": {
        "summary": "Quality",
        "operationId": "quality_api_v1_quality_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Inspect",
        "operationId": "inspect_api_v1_quality_post",
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 120,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Inspection"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users": {
      "get": {
        "summary": "Users",
        "operationId": "users_api_v1_users_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "New User",
        "operationId": "new_user_api_v1_users_post",
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 120,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewAccount"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{user_id}": {
      "post": {
        "summary": "Edit User",
        "operationId": "edit_user_api_v1_users__user_id__post",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 120,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditAccount"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/maintenance/backup": {
      "post": {
        "summary": "Backup",
        "operationId": "backup_api_v1_maintenance_backup_post",
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 120,
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Close": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "Expected Version"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 2,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version",
          "reason"
        ],
        "title": "Close"
      },
      "CustomerReturn": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "customer",
            "title": "Kind"
          },
          "original_id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Original Id"
          },
          "reference": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Reference"
          },
          "quantity": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 1.0,
            "title": "Quantity"
          },
          "expected_version": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "Expected Version"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 2,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "original_id",
          "reference",
          "quantity",
          "expected_version",
          "reason"
        ],
        "title": "CustomerReturn"
      },
...
      "EditAccount": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "warehouse",
              "cashier"
            ],
            "title": "Role"
          },
          "warehouse_ids": {
            "items": {
              "type": "string",
              "maxLength": 120,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 100,
            "title": "Warehouse Ids",
            "default": []
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "expected_version": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "Expected Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "role",
          "active",
          "expected_version"
        ],
        "title": "EditAccount"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Inspection": {
        "properties": {
          "reference": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Reference"
          },
          "return_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Return Id"
          },
          "return_version": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Return Version"
          },
          "warehouse_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Warehouse Id"
          },
          "sku_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Sku Id"
          },
          "good": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "Good",
            "default": 0
          },
          "bad": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "Bad",
            "default": 0
          },
          "expected_version": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "Expected Version"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 2,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reference",
          "expected_version",
          "reason"
        ],
        "title": "Inspection"
      },
...
      "Receive": {
        "properties": {
          "reference": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Reference"
          },
          "expected_version": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "Expected Version"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/ReceiveLine"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Lines"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reference",
          "expected_version",
          "lines"
        ],
        "title": "Receive"
      },
...
      "Review": {
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "approve",
              "reject"
            ],
            "title": "Decision"
          },
          "note": {
            "type": "string",
            "maxLength": 500,
            "minLength": 2,
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "decision",
          "note"
        ],
        "title": "Review"
      },
...
      "SupplierReturn": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "supplier",
            "title": "Kind"
          },
          "original_id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Original Id"
          },
          "reference": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Reference"
          },
          "g": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "G",
            "default": 0
          },
          "q": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "Q",
            "default": 0
          },
          "d": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "D",
            "default": 0
          },
          "expected_version": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "Expected Version"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 2,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "original_id",
          "reference",
          "expected_version",
          "reason"
        ],
        "title": "SupplierReturn"
      },
      "Transfer": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Code"
          },
          "warehouse_id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Warehouse Id"
          },
          "destination_id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Destination Id"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 2,
            "title": "Reason"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/PurchaseLine"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Lines"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "warehouse_id",
          "destination_id",
          "reason",
          "lines"
        ],
        "title": "Transfer"
      },
      "TransferAction": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "Expected Version"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 2,
            "title": "Reason"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/StockVersion"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Lines",
            "default": []
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version",
          "reason"
        ],
        "title": "TransferAction"
      },
...
    }
  }
}